Added Context to all routes and fixes
