+++ src/pages/Home.js
@@ -1,23 +1,23 @@
-import { useEffect, useState } from "react";
-import { getPokemonData, getPokemons, searchPokemon } from "../api";
-import "./App.css";
-import { FavoriteProvider } from "../contexts/favoritesContext";
+import { useEffect, useState, useContext } from "react";
 import Navbar from "../components/Navbar";
 import Pokedex from "../components/Pokedex";
 import Search from "../components/Search";
+import FavoriteContext from "../contexts/favoritesContext";
+import { getPokemonData, getPokemons, searchPokemon } from "../provider/api";
+import "./Home.css";
 
 const favoriteKey = "f";
 
-function App() {
+function Home() {
   const [page, setPage] = useState(0);
   const [totalPages, setTotalPages] = useState(0);
   const [loading, setLoading] = useState(false);
   const [notFound, setNotFound] = useState(false);
   const [pokemons, setPokemons] = useState([]);
-  const [favorites, setFavorites] = useState([]);
-
+  const {favorites, updateFavoritePokemons} = useContext(FavoriteContext);
   // limite dos itens por pagina
   const itensPerPage = 20;
+
   //funcao para puxar os dados da API na url do getPokemonData
   const fetchPokemons = async () => {
     try {
@@ -39,7 +39,6 @@ function App() {
 
   const loadFavoritePokemons = () => {
     const pokemons = JSON.parse(window.localStorage.getItem(favoriteKey)) || []; //precisa definir essa condicao para ele nao dar nulo
-    setFavorites(pokemons);
   };
 
   useEffect(() => {
@@ -49,18 +48,6 @@ function App() {
   useEffect(() => {
     fetchPokemons();
   }, [page]);
-
-  const updateFavoritePokemons = (name) => {
-    const updateFavorites = [...favorites];
-    const favoriteIndex = favorites.indexOf(name);
-    if (favoriteIndex >= 0) {
-      updateFavorites.splice(favoriteIndex, 1);
-    } else {
-      updateFavorites.push(name);
-    }
-    window.localStorage.setItem(favoriteKey, JSON.stringify(updateFavorites));
-    setFavorites(updateFavorites);
-  };
 
   const onSearchHandler = async (pokemon) => {
     if (!pokemon) {
@@ -81,28 +68,21 @@ function App() {
   };
 
   return (
-    <FavoriteProvider
-      value={{
-        favoritePokemons: favorites,
-        updateFavoritePokemons: updateFavoritePokemons,
-      }}
-    >
-      <div>
-        <Navbar />
-        <Search onSearch={onSearchHandler} />
-        {notFound ? (
-          <div className="not-found-text">Sem resultados, tente novamente!</div>
-        ) : (
-          <Pokedex
-            pokemons={pokemons}
-            loading={loading}
-            page={page}
-            setPage={setPage}
-            totalPages={totalPages}
-          />
-        )}
-      </div>
-    </FavoriteProvider>
+    <div>
+      <Navbar />
+      <Search onSearch={onSearchHandler} />
+      {notFound ? (
+        <div className="not-found-text">Sem resultados, tente novamente!</div>
+      ) : (
+        <Pokedex
+          pokemons={pokemons}
+          loading={loading}
+          page={page}
+          setPage={setPage}
+          totalPages={totalPages}
+        />
+      )}
+    </div>
   );
 }
-export default App;
+export default Home;
